refactor(canvas): R2-3 — extract skia_canvas_text.cpp
Phase 4 R2-3 continuation. The SkParagraph-based Canvas2D text surface
relocated from skia_canvas.cpp into skia_canvas_text.cpp (~1,095 lines):
cluster-aware shaping (HarfBuzz + ICU + color-emoji fallback) and every
SkiaCanvas text method that routes through it — set_text_align,
fill_text, fill_text_anchored, fill_text_with_max_width, stroke_text,
fill_text_sdf, measure_text — plus the file-local shape helpers
(shape_with_glyph_fallback, active_typeface_covers_text).

The shared font helpers (get_font_manager, make_stroke_paint,
make_font) carry process-wide font-manager state, so they get a single
external-linkage definition in skia_canvas.cpp (un-static'd) and a
declaration in the existing private skia_canvas_internal.hpp; the
GUARD_CANVAS macro is added there too. Definitions only — public
declarations stay in skia_canvas.hpp.

skia_canvas.cpp: 2,873 → 1,879 (-994). Continues the R2-3 per-feature
split (box_shadow / opacity / shaders / gradients / mask already done).

Verified: pulp-canvas builds clean; pulp-test-canvas (1,375 assertions
/ 43 cases) + pulp-test-canvas2d-shim (167 / 53) pass.

Note: pulp-test-canvas-fonts has one pre-existing failure
(test_canvas_fonts.cpp:303, match_registered_typeface bold-style miss)
that reproduces identically on pristine origin/main — it lives in
bundled_fonts.cpp, untouched by this refactor. Filed separately.
