+++ core/canvas/CMakeLists.txt
@@ -85,6 +85,7 @@ target_link_libraries(pulp-canvas PUBLIC pulp::runtime)
 if(PULP_HAS_SKIA)
     target_sources(pulp-canvas PRIVATE
         src/skia_canvas.cpp
+        src/skia_canvas_text.cpp
         src/skia_canvas_mask.cpp
         src/skia_canvas_gradients.cpp
         src/skia_canvas_box_shadow.cpp
